requirements.txt
requests
python-dotenv
together
jupyter